tars文件更改

+++ src/main/java/com/web/servant/center/orderCloseRequest.java
@@ -26,7 +26,7 @@ public class orderCloseRequest {
 	@TarsStructProperty(order = 6, isRequire = true)
 	public int orderDirection = 0;
 	@TarsStructProperty(order = 7, isRequire = true)
-	public int orderVolume = 0;
+	public double orderVolume = 0D;
 
 	public String getTypeId() {
 		return typeId;
@@ -76,18 +76,18 @@ public class orderCloseRequest {
 		this.orderDirection = orderDirection;
 	}
 
-	public int getOrderVolume() {
+	public double getOrderVolume() {
 		return orderVolume;
 	}
 
-	public void setOrderVolume(int orderVolume) {
+	public void setOrderVolume(double orderVolume) {
 		this.orderVolume = orderVolume;
 	}
 
 	public orderCloseRequest() {
 	}
 
-	public orderCloseRequest(String typeId, int requestId, String brokerId, String userId, String instrumentId, int orderDirection, int orderVolume) {
+	public orderCloseRequest(String typeId, int requestId, String brokerId, String userId, String instrumentId, int orderDirection, double orderVolume) {
 		this.typeId = typeId;
 		this.requestId = requestId;
 		this.brokerId = brokerId;
